address

+++ convert_base.c
@@ -1,7 +1,7 @@
 #include "main.h"
 int handle(unsigned long b, int c, int a)
 {
-	unsigned int l = 0, k, m = 0, n;
+	unsigned long l = 0, k, m = 0, n;
 	char *j, *o = "0123456789ABCDEF", *p = "0123456789abcdef";
 	k = b;
 	if (b == 0)
